requirements.txt
anyio==4.11.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
Flask==3.1.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycryptodome==3.23.0
sniffio==1.3.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
